[mob][photos] Cleanup detections
