--- mobile/lib/services/machine_learning/face_ml/face_detection/face_detection_exceptions.dart
@@ -1,3 +1,0 @@
-class YOLOFaceInterpreterInitializationException implements Exception {}
-
-class YOLOFaceInterpreterRunException implements Exception {}
